Rename types into `sse_types`.
