+++ casper-sse-client/src/sse_types.rs
